requirements.txt
Faker==24.3.0
psycopg2-binary==2.9
